refactor(#742): relocate mat_image.py to core/utils/\n\nMove the PIL image-matting utility from the src/picframe top level to\ncore/utils/, matching the documented Clean Architecture layer layout\n(mat_image is a domain rendering helper, not a top-level module).\n\n- git mv src/picframe/mat_image.py src/picframe/core/utils/mat_image.py\n- Update imports in image_preparer.py, video_frame_extractor.py, test_mat_image.py\n- The ~19 other 'mat_image' grep hits are config keys (viewer.mat_type etc), not imports\n\nGates: ruff, ruff format, mypy, pytest 801 passed.\n\nCloses #742
